// connections/connections/src/app/store/action-type/people-conversation-action.types.ts
export enum PeopleConversationActionTypes {
  GET_PEOPLE_CONVERSATION = '[PEOPLE_CONVERSATION_PAGE] GET PEOPLE_CONVERSATION',
  GET_PEOPLE_CONVERSATION_SUCCESSFUL = '[PEOPLE_CONVERSATION_PAGE] GET PEOPLE_CONVERSATION SUCCESSFUL',
  GET_PEOPLE_CONVERSATION_FAILED = '[PEOPLE_CONVERSATION_PAGE] GET PEOPLE_CONVERSATION FAILED',

  CREATE_PEOPLE_CONVERSATION = '[PEOPLE_CONVERSATION_PAGE] CREATE PEOPLE_CONVERSATION',
  CREATE_PEOPLE_CONVERSATION_SUCCESSFUL = '[PEOPLE_CONVERSATION_PAGE] CREATE PEOPLE_CONVERSATION SUCCESSFUL',
  CREATE_PEOPLE_CONVERSATION_FAILED = '[PEOPLE_CONVERSATION_PAGE] CREATE PEOPLE_CONVERSATION FAILED',

  DELETE_PEOPLE_CONVERSATION = '[PEOPLE_CONVERSATION_PAGE] DELETE PEOPLE_CONVERSATION',
  DELETE_PEOPLE_CONVERSATION_SUCCESSFUL = '[PEOPLE_CONVERSATION_PAGE] DELETE PEOPLE_CONVERSATION SUCCESSFUL',
  DELETE_PEOPLE_CONVERSATION_FAILED = '[PEOPLE_CONVERSATION_PAGE] DELETE PEOPLE_CONVERSATION FAILED'
}